+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to GOBL will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/). See also the [GOBL versions](https://docs.gobl.org/overview/versions) documentation site for more details.
 
+## [vX.XX.X] - XXXX-XX-XX
+
+### Added
+
+- `bill.Invoice`: experimental `ConvertInto` method to convert the invoice's amounts from one currency into another.
+
 ## [v0.81.0] - 2024-07-17
 
 ### Added